read the latest modified cargo index
The previous code used the directory last-modified date to decide
which is the newest index checkout. However, this doesn't account
for the possibility of files within the directory that have been
touched more recently than the directory itself. I don't know how
this situation arises but my guess is it has to do with switching
cargo versions back and forth using rustup.

Ref: #32