Make handling of existing clones more robust
Since by default we e.g., clone projects into `$TEMPDIR`, it can happen
that previous clones get partially reaped. This can lead to existing,
but invalid clones.

Previously we always assumed that a clone either does not exist (and
needs to be fully pulled again), or if it exists is valid (and we can
e.g., simply update it). This is not true if e.g., the git folder was
reaped to such an extent that the clone is broken.

This patch updates our handling to use a slightly modified flow. We
first try to open a repository at the expected location (meaning it both
existed _and_ is valid), or when that fails perform a fresh clone
(removing any existing clone folder). This allows us to handle a wider
range of failure scenarios (e.g., git directory in clone partially
reaped; some permission changes on clone like clone being not readable).